Renames braggdiskdetection -> diskdetection
