+++ py4DSTEM/process/__init__.py
@@ -1,7 +1,7 @@
 from . import utils
 from . import fit
 from . import preprocess
-from . import braggdiskdetection
+from . import diskdetection
 from . import calibration
 from . import latticevectors
 from . import classification
