+++ py4DSTEM/process/diskdetection/diskdetection_parallel.py
@@ -9,7 +9,7 @@ import numpy as np
 
 # local
 import py4DSTEM
-from py4DSTEM.process.braggdiskdetection import PointListArray
+from py4DSTEM.process.diskdetection import PointListArray
 
 
 def _find_Bragg_disks_single_DP_FK(DP, probe_kernel_FT,
