+++ py4DSTEM/process/diskdetection/probetemplate.py
